+++ classes/warrior.lua
@@ -1,11 +1,11 @@
 local AddonName, SAO = ...
 
-local function registerAuras(self)
+local function registerClass(self)
     self:RegisterAura("bloodsurge", 0, 46916, "blood_surge", "Top", 1, 255, 255, 255, true);
     self:RegisterAura("sudden_death", 0, 52437, "sudden_death", "Left + Right (Flipped)", 1, 255, 255, 255, true);
     self:RegisterAura("sword_and_board", 0, 50227, "sword_and_board", "Left + Right (Flipped)", 1, 255, 255, 255, true);
 end
 
 SAO.Class["WARRIOR"] = {
-    ["Register"] = registerAuras,
+    ["Register"] = registerClass,
 }